feat(settings-profile): replace use of vuex account state with iridium state
