+++ store/accounts/actions.ts
@@ -71,7 +71,7 @@ export default {
         pin,
       )
 
-      await commit('setPhrase', decryptedPhrase)
+      commit('setPhrase', decryptedPhrase)
     }
 
     commit('unlock', pin)
@@ -91,7 +91,7 @@ export default {
       throw new Error(AccountsError.INVALID_PIN)
     }
 
-    await commit('setAdapter', 'Solana')
+    commit('setAdapter', 'Solana')
     const $BlockchainClient: BlockchainClient = BlockchainClient.getInstance()
     $BlockchainClient.setAdapter(new SolanaAdapter())
 
@@ -102,7 +102,7 @@ export default {
       throw new Error(AccountsError.UNABLE_TO_CREATE_MNEMONIC)
     }
 
-    await commit('setPhrase', userWallet.mnemonic)
+    commit('setPhrase', userWallet.mnemonic)
 
     const { pinHash } = state
     const entropyMessage = IdentityManager.generateEntropyMessage(
@@ -140,7 +140,7 @@ export default {
 
     const encryptedPhrase = await Crypto.encryptWithPassword(mnemonic, pin)
 
-    await commit('setEncryptedPhrase', encryptedPhrase)
+    commit('setEncryptedPhrase', encryptedPhrase)
   },
   /**
    * @method loadAccount
@@ -284,22 +284,6 @@ export default {
     dispatch('startup', walletAccount)
   },
 
-  async updateUser({ commit }: ActionsArguments<AccountsState>, details: User) {
-    // update iridium state
-    const detailsKeys = Object.keys(details) as (keyof User)[]
-    await Promise.all(
-      detailsKeys.map(async (key) => {
-        if (Object.prototype.hasOwnProperty.call(details, key)) {
-          const value = details[key as keyof User]
-          await iridium.profile?.set(`/${key}`, value)
-        }
-      }),
-    )
-
-    // update vuex state
-    commit('updateUserDetails', details)
-  },
-
   /**
    * @method updateProfilePhoto
    * @description update profile photo of the user on the Solana blockchain
@@ -335,7 +319,7 @@ export default {
   ) {
     // Initialize crypto engine
     const $Crypto: Crypto = Vue.prototype.$Crypto
-    await $Crypto.init(userAccount)
+    $Crypto.init(userAccount)
   },
 
   async startup({
